+++ tests/Paket.Tests/InstallModel/Xml/Fantomas.fs
@@ -21,7 +21,7 @@ let ``should generate Xml for Fantomas 1.5``() =
             [ @"..\Fantomas\lib\FantomasLib.dll" 
               @"..\Fantomas\lib\FSharp.Core.dll" 
               @"..\Fantomas\lib\Fantomas.exe" ],
-              Nuspec.Explicit ["FantomasLib.dll"])
+              Nuspec.Explicit ["FantomasLib.dll"]).FilterFallbacks()
     
     let chooseNode = ProjectFile.Load("./ProjectFile/TestData/Empty.fsprojtest").Value.GenerateXml(model)
     chooseNode.OuterXml