+++ tests/Paket.Tests/InstallModel/Xml/SystemNetHttp.fs
@@ -22,7 +22,28 @@ let expected = """
         <ItemGroup />
       </When>
       <When Condition="$(TargetFrameworkVersion) == 'v4.0' And $(TargetFrameworkProfile) == 'Client'">
-        <ItemGroup />
+        <ItemGroup>
+          <Reference Include="System.Net.Http.Extensions">
+            <HintPath>..\..\..\Microsoft.Net.Http\lib\net40\System.Net.Http.Extensions.dll</HintPath>
+            <Private>True</Private>
+            <Paket>True</Paket>
+          </Reference>
+          <Reference Include="System.Net.Http.Primitives">
+            <HintPath>..\..\..\Microsoft.Net.Http\lib\net40\System.Net.Http.Primitives.dll</HintPath>
+            <Private>True</Private>
+            <Paket>True</Paket>
+          </Reference>
+          <Reference Include="System.Net.Http.WebRequest">
+            <HintPath>..\..\..\Microsoft.Net.Http\lib\net40\System.Net.Http.WebRequest.dll</HintPath>
+            <Private>True</Private>
+            <Paket>True</Paket>
+          </Reference>
+          <Reference Include="System.Net.Http">
+            <HintPath>..\..\..\Microsoft.Net.Http\lib\net40\System.Net.Http.dll</HintPath>
+            <Private>True</Private>
+            <Paket>True</Paket>
+          </Reference>
+        </ItemGroup>
       </When>
       <When Condition="$(TargetFrameworkVersion) == 'v4.0'">
         <ItemGroup>
@@ -272,7 +293,7 @@ let ``should generate Xml for System.Net.Http 2.2.8``() =
               
               @"..\Microsoft.Net.Http\lib\wpa81\System.Net.Http.Extensions.dll" 
               @"..\Microsoft.Net.Http\lib\wpa81\System.Net.Http.Primitives.dll" ],
-              Nuspec.All)
+              Nuspec.All).FilterFallbacks()
 
     let chooseNode = ProjectFile.Load("./ProjectFile/TestData/Empty.fsprojtest").Value.GenerateXml(model)
     chooseNode.OuterXml