+++ tests/Paket.Tests/NugetVersionRangeSerializerSpecs.fs
@@ -1,0 +1,43 @@
+﻿module Paket.NugetVersionRangeSerializerSpecs
+
+open Paket
+open NUnit.Framework
+open FsUnit
+
+[<Test>]
+let ``can detect latest version``() = 
+    NugetVersionRangeParser.parse "" |> shouldEqual VersionRequirement.AllReleases
+
+let format(versionRange:VersionRange) = NugetVersionRangeParser.format versionRange
+
+[<Test>]
+let ``can detect specific version``() = 
+    VersionRange.Specific (SemVer.Parse "2.2") |> format |> shouldEqual "[2.2]"
+    VersionRange.Specific (SemVer.Parse "1.2") |> format |> shouldEqual "[1.2]"
+
+[<Test>]
+let ``can detect minimum version``() = 
+    VersionRange.Minimum (SemVer.Parse "2.2") |> format |> shouldEqual "2.2"
+    VersionRange.Minimum (SemVer.Parse "1.2") |> format |> shouldEqual "1.2"
+
+[<Test>]
+let ``can detect greater than version``() = 
+    VersionRange.GreaterThan (SemVer.Parse "2.2") |> format |> shouldEqual "(2.2,)"
+    VersionRange.GreaterThan (SemVer.Parse "1.2") |> format |> shouldEqual "(1.2,)"
+
+[<Test>]
+let ``can detect maximum version``() = 
+    VersionRange.Maximum (SemVer.Parse "2.2") |> format |> shouldEqual "(,2.2]"
+    VersionRange.Maximum (SemVer.Parse "1.2") |> format |> shouldEqual "(,1.2]"
+
+[<Test>]
+let ``can detect less than version``() = 
+    VersionRange.LessThan (SemVer.Parse "2.2") |> format |> shouldEqual "(,2.2)"
+    VersionRange.LessThan (SemVer.Parse "1.2") |> format |> shouldEqual "(,1.2)"
+
+[<Test>]
+let ``can detect range version``() = 
+    VersionRange.Range(VersionRangeBound.Excluding, SemVer.Parse "2.2", SemVer.Parse "3", VersionRangeBound.Excluding) |> format |> shouldEqual "(2.2,3)" 
+    VersionRange.Range(VersionRangeBound.Excluding, SemVer.Parse "2.2", SemVer.Parse "3", VersionRangeBound.Including) |> format |> shouldEqual "(2.2,3]" 
+    VersionRange.Range(VersionRangeBound.Including, SemVer.Parse "2.2", SemVer.Parse "3", VersionRangeBound.Excluding) |> format |> shouldEqual "[2.2,3)" 
+    VersionRange.Range(VersionRangeBound.Including, SemVer.Parse "2.2", SemVer.Parse "3", VersionRangeBound.Including) |> format |> shouldEqual "[2.2,3]"